add ARMv8 assembler and CE support
Addition includes Linux kernel support. Assembler implementation measures
7.5 seconds for KMAC DRNG 1GB vs 10 seconds for C.

+++ hash/src/asm/ARMv7A-NEON/KeccakP-1600-SnP.h
@@ -21,6 +21,8 @@ Please refer to SnP-documentation.h for more details.
 #ifndef _KeccakP_1600_SnP_h_
 #define _KeccakP_1600_SnP_h_
 
+#include "ext_headers.h"
+
 #define KeccakP1600_stateSizeInBytes    200
 #define KeccakP1600_stateAlignment      32
 
@@ -41,7 +43,7 @@ void KeccakP1600_Permute_Nrounds(void *state, unsigned int nrounds);
 void KeccakP1600_Permute_12rounds(void *state);
 void KeccakP1600_Permute_24rounds(void *state);
 void KeccakP1600_ExtractBytes(const void *state, unsigned char *data,
-			      unsigned int offset, unsigned int length);
+			      size_t offset, size_t length);
 void KeccakP1600_ExtractAndAddBytes(const void *state,
 				    const unsigned char *input,
 				    unsigned char *output, unsigned int offset, unsigned int length);
