+++ hash/src/asm/ARMv8A/KeccakP-1600-SnP.h
@@ -12,6 +12,8 @@ Please refer to SnP-documentation.h for more details.
 #ifndef _KeccakP_1600_SnP_h_
 #define _KeccakP_1600_SnP_h_
 
+#include "ext_headers.h"
+
 #define KeccakP1600_implementation      "64-bit optimized ARMv8a assembler implementation"
 #define KeccakP1600_stateSizeInBytes    200
 #define KeccakP1600_stateAlignment      64
@@ -28,7 +30,7 @@ void KeccakP1600_Permute_Nrounds(void *state, unsigned int nrounds);
 void KeccakP1600_Permute_12rounds(void *state);
 void KeccakP1600_Permute_24rounds(void *state);
 void KeccakP1600_ExtractBytes(const void *state, unsigned char *data,
-			      unsigned int offset, unsigned int length);
+			      size_t offset, size_t length);
 void KeccakP1600_ExtractAndAddBytes(const void *state,
 				    const unsigned char *input,
 				    unsigned char *output, unsigned int offset, unsigned int length);
